Make MuonGun compile cleanly with a reasonable subset of clang's -Weverything.
Disabled warnings:
global-constructors, exit-time-destructors (unavoidable)
documentation-unknown-command (clang knows doxygen, but not well enough)
padded (I don't care about the memory layout of I3Modules)
float-equal (it's not safe in general, but we use this for object equivalence all the time)
c99-extensions (variadic macros are nice)


git-svn-id: http://code.icecube.wisc.edu/svn/projects/MuonGun/trunk@128654 16731396-06f5-0310-8873-f7f720988828

+++ private/MuonGun/Flux.cxx
@@ -55,13 +55,13 @@ SplineFlux::SplineFlux(const std::string &singles, const std::string &bundles)
     : singles_(singles), bundles_(bundles)
 {
 	SetMinMultiplicity(1);
-	SetMaxMultiplicity(bundles_.GetExtents(2).second);
+	SetMaxMultiplicity(static_cast<unsigned>(bundles_.GetExtents(2).second));
 }
 
 double
 SplineFlux::GetLog(double depth, double cos_theta, unsigned multiplicity) const
 {
-	double coords[3] = {cos_theta, depth, multiplicity};
+	double coords[3] = {cos_theta, depth, static_cast<double>(multiplicity)};
 	double logflux;
 	
 	if (multiplicity < GetMinMultiplicity() || multiplicity > GetMaxMultiplicity())
